requirements.txt
requests==2.31.0
py-cord==2.4.1
python-dotenv==1.0.0
flask
gradio_client